+++ packages/server/src/orchestrator/registry.ts
@@ -110,12 +110,12 @@ export async function pruneStaleNodes(db: FairygitMotherDb, timeoutMs: number): 
 
 export async function getActiveNodeCount(db: FairygitMotherDb): Promise<number> {
 	const result = (
-		await db.select({ count: sql<number>`count(*)` }).from(nodes).where(ne(nodes.status, "offline"))
+		await db.select({ count: sql<number>`count(*)::int` }).from(nodes).where(ne(nodes.status, "offline"))
 	)[0];
 	return result?.count ?? 0;
 }
 
 export async function getTotalNodeCount(db: FairygitMotherDb): Promise<number> {
-	const result = (await db.select({ count: sql<number>`count(*)` }).from(nodes))[0];
+	const result = (await db.select({ count: sql<number>`count(*)::int` }).from(nodes))[0];
 	return result?.count ?? 0;
 }
